# openapi.yaml
openapi: 3.1.0
info:
  title: OCE Middleware
  version: "0.2.1"

servers:
  - url: /   https://oce-middleware.onrender.com

components:
  securitySchemes:
    ActionKey:
      type: apiKey
      in: header
      name: X-Action-Key
  schemas:
    SessionCtx:
      type: object
      required: [project_id]
      properties:
        project_id: { type: string }
        user_id: { type: string }
        message_id: { type: string }
        mcda: { type: object, additionalProperties: true }
        risk: { type: object, additionalProperties: true }
        meta: { type: object, additionalProperties: true }

    RunRequest:
      type: object
      required: [user_text, session_ctx]
      properties:
        user_text: { type: string }
        session_ctx:
          $ref: '#/components/schemas/SessionCtx'

    HealthResponse:
      type: object
      required: [status, middleware]
      properties:
        status: { type: string, enum: [ok, degraded] }
        middleware: { type: string }
        oce: { type: object, additionalProperties: true }
        error: { type: string }

    WarmupResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [warmed] }
        public: { type: boolean }
        warning: { type: string }

    OCEBridgeResponse:
      type: object
      description: OCE upstream payload plus middleware-added assistant_text.
      properties:
        text: { type: string }
        json_summary: { type: object, additionalProperties: true }
        telemetry: { type: object, additionalProperties: true }
        assistant_text:
          type: string
          description: Human-readable short answer produced by middleware.
      additionalProperties: true

paths:
  /health:
    get:
      operationId: health
      summary: Lightweight middleware liveness
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /health/deep:
    get:
      operationId: healthDeep
      summary: Check upstream OCE health (manual only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /warmup:
    post:
      operationId: warmup
      summary: Wake middleware; optional ping to OCE
      security:
        - ActionKey: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Warmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarmupResponse'
        "401": { description: Unauthorized }

  /bridge/run:
    post:
      operationId: bridgeRun
      summary: Route a user request to OCE with safe shaping; returns assistant_text
      security:
        - ActionKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
      responses:
        "200":
          description: OCE response with assistant_text
          headers:
            X-Request-ID: { schema: { type: string } }
            X-Memory-Check: { schema: { type: string, enum: ["ok","fail"] } }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OCEBridgeResponse'
        "401": { description: Unauthorized }
        "429": { description: Rate limited }
        "502": { description: Upstream error }